Implemented Missing Entities
Forgot include the current level text that is present in the Canva design, the text that is situated between the Play and Quit button.

Also coded for that  when the Play button it pressed, the level goes up by 1 and the play button is disabled.

The timer also begins counting down when the Play button is pressed, further when the number becomes 5 and less the text turns red to visualize to the Player urgency.