Final Addition: Setup Implementation
Implemented the mechanism the allows the wanted shape to be displayed at the top middle of the screen within the infoPanel, upon pressing the PLAY button, the wanted shape will appear and the four shapes also will appear in the gamePanel window.

Implemented proper correspondence so that the wanted shape displayed in the infoPanel, is the ACTUAL wanted shape, so when a player clicks the correct shape within the gamePanel, they get  rewarded with 100 points and  + 5 seconds, on the contrary, the player will lose 10 seconds.

In this current implementation, you'd need to press the Play button again for a new wanted shape to appear.