Implement New Level Type: Mimic
In these levels all the unwanted shapes will take on the colour of the wanted shape, mimicking him. The general form of the level is layout as a gridLevel, and given that I decided to merge those two levels into one, including a boolean flag to turn mimicry on and off easily.

Since this also made changed to the populateGrid function, needed the wantedColor in its parameters as well as the mimic flag, other levels got affected such as Motion, so now motion levels also have the ability to implement mimicry or not

And since Flicker used the gridLevel as a layout to perform it's flicker effect, it now uses the mimicLevel to do so, also allowing flicker levels the ability to mimic or not

The addition create the possibility for so many other levels!