Refactoring: GamePanel
Did some important refactoring for the GamePanel class, previously it was violating Single Responsibility Principle (SRP) by managing the what goes on within the panel as well as the levels involved.

Decided to make a new class, levelManager, which is responsible for generating the different levels, communicating with the gamePanel, and the gamePanel is responsible for placing those generated level within it dimensions.