Implemented New Level Type: Relay
Through this level the requirement of having something interesting happen when game objects collide is satisfied, through color swiping! When a shape collides with another whoever collides first they will take the color of the collided shape that includes the wanted shape! They're all playing a game of relay but a more chaotic version

I combined this logic with the motionLevel since it also uses a grid and the only difference is whether collision is enabled or not

This shall be the final level type I implement, the rest of the development will just be making a complete game with different levels where I just manipulate these level types